+++ src/cryptonote_core/tx_pool.cpp
@@ -411,7 +411,7 @@ namespace cryptonote
                                           << ",  kei_image_set.size()=" << kei_image_set.size() << ENDL << "txin.k_image=" << txin.k_image << ENDL
                                           << "tx_id=" << id );
       auto ins_res = kei_image_set.insert(id);
-      CHECK_AND_ASSERT_MES(ins_res.second, false, "internal error: try to insert duplicate iterator in key_image set");
+      CHECK_AND_WARN_MES(ins_res.second, false, "Try to insert duplicate iterator in key_image set");
     }
     ++m_cookie;
     return true;